admin_panel/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
django-quill-editor==0.1.40
gunicorn==20.0.4
psycopg2-binary==2.9
python-dotenv==0.19.2
sqlparse==0.4.2
